Refactor into multiple modules
