+++ Cargo.toml
@@ -36,6 +36,7 @@ event-listener = "2.5.3"
 line-straddler = "0.2.0"
 piet = { version = "0.6.2", default-features = false }
 rayon-core = { version = "1.11.0", optional = true }
+tinyvec = { version = "1.6.0", default-features = false, features = ["alloc"] }
 tracing = { version = "0.1.37", default-features = false, optional = true }
 yazi = { version = "0.1.6", optional = true }
